Fix TopDownPlayer speed

+++ Math/Vector2i.h
@@ -235,17 +235,25 @@
             ////////////////////////////////////////////////////////////////////
             //  Get the dot product of this vector and another                //
             ////////////////////////////////////////////////////////////////////
-            inline int32_t dotProduct(const Vector2i& v) const
+            inline int64_t dotProduct(const Vector2i& v) const
             {
-                return ((vec[0]*v.vec[0]) + (vec[1]*v.vec[1]));
+                return (
+                    (static_cast<int64_t>(vec[0])*
+                    static_cast<int64_t>(v.vec[0]))+
+                    (static_cast<int64_t>(vec[1])*
+                    static_cast<int64_t>(v.vec[1]))
+                );
             }
 
             ////////////////////////////////////////////////////////////////////
             //  Get Vector2i length                                           //
             ////////////////////////////////////////////////////////////////////
-            inline int32_t length() const
+            inline int64_t length() const
             {
-                return Math::sqrt((vec[0]*vec[0]) + (vec[1]*vec[1]));
+                return Math::sqrt(
+                    (static_cast<int64_t>(vec[0])*static_cast<int64_t>(vec[0]))+
+                    (static_cast<int64_t>(vec[1])*static_cast<int64_t>(vec[1]))
+                );
             }
 
             ////////////////////////////////////////////////////////////////////
@@ -253,12 +261,15 @@
             ////////////////////////////////////////////////////////////////////
             inline void normalize()
             {
-                int32_t len = length();
+                int64_t len = length();
                 if (len != 0)
                 {
-                    // Normalize between 0 and Math::OneInt
-                    vec[0] = (vec[0] << Math::OneIntShift) / len;
-                    vec[1] = (vec[1] << Math::OneIntShift) / len;
+                    vec[0] = static_cast<int32_t>(
+                        (static_cast<int64_t>(vec[0])<<Math::OneIntShift) / len
+                    );
+                    vec[1] = static_cast<int32_t>(
+                        (static_cast<int64_t>(vec[1])<<Math::OneIntShift) / len
+                    );
                 }
             }
