+++ Math/Vector4i.h
@@ -334,22 +334,30 @@
             ////////////////////////////////////////////////////////////////////
             //  Get the dot product of this vector and another                //
             ////////////////////////////////////////////////////////////////////
-            inline int32_t dotProduct(const Vector4i& v) const
+            inline int64_t dotProduct(const Vector4i& v) const
             {
                 return (
-                    (vec[0]*v.vec[0]) + (vec[1]*v.vec[1]) +
-                    (vec[2]*v.vec[2]) + (vec[3]*v.vec[3])
+                    (static_cast<int64_t>(vec[0])*
+                    static_cast<int64_t>(v.vec[0]))+
+                    (static_cast<int64_t>(vec[1])*
+                    static_cast<int64_t>(v.vec[1]))+
+                    (static_cast<int64_t>(vec[2])*
+                    static_cast<int64_t>(v.vec[2]))+
+                    (static_cast<int64_t>(vec[3])*
+                    static_cast<int64_t>(v.vec[3]))
                 );
             }
 
             ////////////////////////////////////////////////////////////////////
             //  Get Vector4i length                                           //
             ////////////////////////////////////////////////////////////////////
-            inline int32_t length() const
+            inline int64_t length() const
             {
                 return Math::sqrt(
-                    (vec[0]*vec[0]) + (vec[1]*vec[1]) +
-                    (vec[2]*vec[2]) + (vec[3]*vec[3])
+                    (static_cast<int64_t>(vec[0])*static_cast<int64_t>(vec[0]))+
+                    (static_cast<int64_t>(vec[1])*static_cast<int64_t>(vec[1]))+
+                    (static_cast<int64_t>(vec[2])*static_cast<int64_t>(vec[2]))+
+                    (static_cast<int64_t>(vec[3])*static_cast<int64_t>(vec[3]))
                 );
             }
 
@@ -358,14 +366,21 @@
             ////////////////////////////////////////////////////////////////////
             inline void normalize()
             {
-                int32_t len = length();
-                if (len != 0.0f)
+                int64_t len = length();
+                if (len != 0)
                 {
-                    // Normalize between 0 and Math::OneInt
-                    vec[0] = (vec[0] << Math::OneIntShift) / len;
-                    vec[1] = (vec[1] << Math::OneIntShift) / len;
-                    vec[2] = (vec[2] << Math::OneIntShift) / len;
-                    vec[3] = (vec[3] << Math::OneIntShift) / len;
+                    vec[0] = static_cast<int32_t>(
+                        (static_cast<int64_t>(vec[0])<<Math::OneIntShift) / len
+                    );
+                    vec[1] = static_cast<int32_t>(
+                        (static_cast<int64_t>(vec[1])<<Math::OneIntShift) / len
+                    );
+                    vec[2] = static_cast<int32_t>(
+                        (static_cast<int64_t>(vec[2])<<Math::OneIntShift) / len
+                    );
+                    vec[3] = static_cast<int32_t>(
+                        (static_cast<int64_t>(vec[3])<<Math::OneIntShift) / len
+                    );
                 }
             }
